LU-3546 lod: cancel update log after all committed
Cancel the log cookies after all of updates have been
committed to disk.

In DNE2, the distributed operation will write update log in all
of involve MDTs, when these the updates are committed on all of
MDT, these update log will be cancelled. If one of these MDTs
restart during the operation, it will replay the updates based
on these update log.

The distribute operation will be represent as distribute transaction
(top_thandle), when the distribute transaction is commited
(see lod_sub_txn_commit_cb()), the top_thandle will be added to the
lod_commit_list, and the log cancel thread will be wakeup and walk
through the commit list and cancel the update logs for these transaction.

During the cancellation, update logs on local master MDT will be cancelled
first, after it is committed the disk, it will then cancel the update log
in other MDTs.  The transaction can only be destroyed after all of update
logs are cancelled.

To add transaction commit callback during the llog cancellation, it also
need to add thandle as the parameter to the API of llog cookie cancellation.
for example llog_cat_cancel_records() llog_cancel_rec().

Change-Id: I3023f95515b2c15307479f7f85c3f13926d4bcc9
Signed-off-by: Wang Di <di.wang@intel.com>