LU-2182 llapi: Expand layout api to verify against a target file system
Besides validating the layout users create against the limits of
Lustre add additional functions that can validate the layout
against the target file system. To do this two function are added.
The function llapi_layout_pool_absent can be used to ensure the
layout's pool is actually present on the file system. The second
function llapi_layout_ost_index_valid ensures that the user
specified OST index supplied matches to a real OST on the file
system. This method also test if the OST index belongs to the
layout's pool if one is present.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: Ie2eb1538881d3e6b070c45d50a4a8c3dd87052c7