LU-5424 tests: exclude sanity-sec test 4
The patch to fix LU-5424 causing sanity-sec test_4 to fail was
landed without test_4 actually enabled in the test environment.
Need to disable this in sanity-sec ALWAYS_EXCEPT again, re-enable
it in the testing environment, then revert this patch to see if
the test is passing or if it needs more fixing.

Change-Id: I0759f4ab7022bc072bb2480eaba0b19ff0f0551e