LU-5570 lnet: check router aliveness timestamp
In current LNet, peer aliveness timestamp is only recorded for peers
on routers, this patch will change this and also record aliveness of
routers on regular nodes, so lnet can check aliveness timestamp of
routers before sending message, avoid to choose router which has old
aliveness timestamp and can be potentially dead or congested.

This is an experimental patch, not for production yet.

Signed-off-by: Liang Zhen <liang.zhen@intel.com>
Change-Id: I305ba684ced231a87656458382966aeb8ff7ba48