LU-5530 mdt: Force resent due to large replies.
This is currently a test only patch to force resending
of requests as if we have large striping in play which is common
on larger sites, but not in our testing.

Eventually this will be a fail_loc, I think, but RIGHT NOW THIS
IS NOT INTENDED FOR LANDING.

This also causes a bunch of client reconnections in racer for some
unknown reason in my testing.

Change-Id: Ibfc131c72b6f810fc7380035268ed47f613989fb
Signed-off-by: Oleg Drokin <oleg.drokin@intel.com>