LU-3319 procfs: provide framework for seq_file handling
In linux 3.10 kernels and above the proc file system interface
changed. To handle this change we need to move from the current
method of using proc_read and proc_write function pointers to
using the seq_file. This patch provides the base to enable moving
all the lustre proc handling gradually over to this new method
of proc data handling.

This patch is back-ported from the following one:
Lustre-commit: 3270bfc2370884933628f95122da00d430db6072
Lustre-change: http://review.whamcloud.com/7135

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I8d03093130d2d8ad629f782271e872507d5de319