LU-3319 procfs: provide framework for seq_file handling
In linux 3.10 kernels and above the proc file system interface
changed. To handle this change we need to move from the current
method of using proc_read and proc_write function pointers to
using the seq_file. This patch provides the base to enable moving
all the lustre proc handling gradually over to this new method
of proc data handling.

This patch is back-ported from the following one:
Lustre-commit: 3270bfc2370884933628f95122da00d430db6072
Lustre-change: http://review.whamcloud.com/7135

LU-4287 autoconf: Improve test for proc write and read functions

For newer kernels the proc api has changed to using only
seq_files. To detect this I tested the presence of PDE_DATA
to determine if it was a newer kernel. Now in RHEL6.5 this
macro has been backported but the rest of the changes have
not so it has broken the build. This patch moves from
testing PDE_DATA to testing if you can call write_proc_t
in struct proc_dir_entry. In newer kernels that function
is missing.

This patch is back-ported from the following one:
Lustre-commit: 1e60ce443bdb4d1211758b0e53aea597bbdbc22d
Lustre-change: http://review.whamcloud.com/8482

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Jian Yu <jian.yu@intel.com>
Change-Id: I8d03093130d2d8ad629f782271e872507d5de319