LU-5903 build: Make "make rpm" use "make src" output
We rework the rpm and srpm Make targets.  The srpm
target generates the source rpm package much as it did
before, but it puts the results in a known temporary
rpm build tree.  This allows the "make rpm" target to
simply depend on the srpm target, and then build the
binary rpms from the source rpm.

Previously the rpm target _also_ generated a source rpm
package, which could be redudant, and failed to verify that
the source rpm package was itself buildable.

Further the patch moves the construction of the rpmbuild
command line options out of the makefile and into the m4
files.  Scripting is much easier in the m4 where the contents
are mostly shell anyway, and it also removes a somewhat
confusing layer of indirection from the top of the
autoMakefile.am file.

Change-Id: Id0612ade9f170d84518f01d4e8de654a7ba8af88
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>