LU-5820 lfsck: use multiple namespace LFSCK trace files
The namespace LFSCK uses trace file to record the FID of the object
that has multiple hard links, or has remote name entry, or contains
some uncertain inconsistency, and so on. Only single namespace LFSCK
trace file may be not efficient, especially when there are millions
of FIDs to be recorded. So use multiple namespace LFSCK trace files
and per trace file based semaphore to control the concurrent access
of the trace file.

This patch also drops some repeated FID recording in the namespace
LFSCK trace file. Related FID should has been recorded in the trace
file via lfsck_namespace_exec_oit(), it is unnecessary to do that
again when scans the directory.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Iec27c52b21789dbde1e4c1153f61162f028ceac3