LU-5141 hsm: Only regular files should be archived
It is currently possible to ask lfs to hsm_archive a directory,
although this doesn't appear to make any sense, and the posix
copytool rejects it.

Ideally this should be caught early and not being forwarded to
the copytool at all.

So adding regular file check in
lustre/utils/lfs.c
which will report error if the file provided for hsm_archive
is not a regular file

Signed-off-by: vinayakswami hariharmath <vinayakswami.hariharmath@clogeny.com>
Change-Id: I07175aae4a4605cce74ccc2491e5e185cccdd912