LU-5231 hsm: incorrect progress reported by lfs hsm_action
When hsm_action command is executed on a file on which
hsm archieve or hsm restore was going on, the progress report show
was wrong.

The issue is hsm_action gets the progress report (how much data
moved) from kernel(through ioctl) and the size will be updated
through ioctl from copy tool.
Basically copy tool is sending amount of data moved
(progress report) properly but the data which is sent from copy tool
to kernel was getting added every time with previous value rather
than simply assigning the value.
changes made: Add operation is taken out from the current code
fix is in: lustre/mdt/mdt_hsm_cdt_requests.c

The issue is not only with 5GB file but with all the files. For small
files it was not observed because interval for ioctl on copy tool was
long(30 sec),
changes made: reduced the time interval to 10 sec to get some good,
frequent progress report on client side
fix is in: lustre/utils/lhsmtool_posix.c

also final update about the progress was not used to sent to kernel.
changes made: Once the copy operation is finished, final ioctl is
sent to kernel
fix is in: lustre/utils/lhsmtool_posix.c

Signed-off-by: vinayak hariharmath <vinayakswami.hariharmath@clogeny.com>
Change-Id: I368715db5b7ed67a80005b0cfd5c64ea1ee89d16