LU-5644 clio: yield CPU when cl_lock isn't pruned immediately
cl_locks_prune() will not delete a cl_lock when its cll_holds value
is greater than 0. In this case, the function leaves the cl_lock on
the object's coh_lock list and iterates in a tight loop examining
the same lock until all holds are released. It does not yield the
CPU while looping. This causes CPU stalls and heartbeat faults if
the holds are not released quickly enough. As an example,
osc_lock_enqueue() increments the hold count before issuing its
LDLM_ENQUEUE request. Thus cl_locks_prune can block the CPU for the
time it takes to complete or timeout an RPC.

cl_locks_prune() already calls l_wait_event() for cll_users. It also
needs to wait for cll_holds to become 0.

Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: If64d870d813c64bf96234f63ddf3c75df5672f82