LU-5474 hsm: Check HSM status using hsm_state
Many tests in sanity-hsm look at the hsm/actions on
the MDT to check that HSM commands completed. Now we
check the hsm_state of the file on the client.

This patch is intended to collect data on
sanity-hsm test_90 failures and if it fixes
test 90.

Test-Parameters: fortestonly \
testlist=sanity-hsm,sanity-hsm,sanity-hsm,sanity-hsm,sanity-hsm,sanity-hsm,sanity-hsm
Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I0beee30b681d4b80f23d33cb42ff5b2944fc21d1