LU-0000 osd: cache LOVEA in osd-ldiskfs
there is an idea that EA fetching is an expensive operation which
affects creation rate. it's interesting to know whether we could
move LOV EA caching down to OSD and solve the outstanding with
caching LOV EA in LOD which is hard to revalidate when an object
is being accessed via OUT.

the primary purpose of the patch is to verify the approch is
functionally correct. then we'll be able to benchmark it.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: Ie7925c94cc2147c3650af18c6118523d48624ebe