LU-5727 ldlm: revert the changes for lock canceling policy
The changes for LRU lock policy was introduced by commit 0a6c6fcd4,
where I was trying to revise the policy to pick locks for canceling.

However, this caused two problems as mentioned in LU-5727. The first
problem is that the lock can only be picked for canceling only if
the number of LRU locks is over preset LRU number AND it's aged; the
second problem is that mdc_cancel_weight() tends to not cancel OPEN
locks, therefore open locks can be kept forever and finally exhausts
memory on the MDT side.

This patch will revert the changes related to LRU policy revision.

Lustre-commit: e8812867e8c6c9eb62174b370cb818985e3d2180
Lustre-change: http://review.whamcloud.com/12448

Signed-off-by: Niu Yawei <yawei.niu@intel.com>
Signed-off-by: Jinshan Xiong <jinshan.xiong@intel.com>
Change-Id: Ic04f894c8bf54dbbdfb0e3bada5f6e4a367711a5