LU-5817 clio: Remove cg->cg_gid assertion
When a group lock with GID=0 is released (put_grouplock is
called), an assertion in cl_put_grouplock is hit.

There are two assertions at the start of cl_put_grouplock:
LASSERT(cg->cg_env);
LASSERT(cg->cg_gid);

cg_env is a pointer, so that assertion is valid, but cg_gid
is an integer so that is not a valid assertion.
There are no other checks to make sure the GID is not zero,
so a zero a GID is otherwise valid.

Since there doesn't appear to be any reason a '0' GID
shouldn't be allowed, we can remove the assertion.

Change-Id: I56e58791742809da5353a4d8dfbf3b80a22f3814
Signed-off-by: Patrick Farrell <paf@cray.com>