LU-3538 dne: add COS support
if COS is enabled, cross-MDT lock can eliminate transaction
dependency naturally, because it's a client lock, which will be
canceled only when there is no reader or writer, which means reply
state has been handled and transaction has been committed.

and for the same reason, mdt_save_lock() doesn't need to commit
async for cross-MDT lock, so the 'sync' flag is useless for
cross-MDT lock.

client cookie is also useless for cross-MDT lock, because any
lock contention will go to the slave MDT.

This patch just saves cross-MDT lock to make COS work.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I91928d097cbb26bd1e1089c3f8851ac6a6440a69