LU-3538 dne: Commit-on-Sharing for DNE
This patch contains three parts:
1. Sync-on-Cancel for cross-MDT lock, which eleminates dependency
   between locks and existed PW/EX cross-MDT lock.
2. enable Commit-on-Sharing for DNE, PW/EX locks will be converted
   to COS locks, but by default they are ignored, when operation
   finds itself a distributed transaction, it will lock with
   LDLM_FL_COS_INCOMPAT flag to check against existed COS locks.
   This will eliminate dependency between locks and COS locks for
   distributed transaction.
3. striped directory creation needs to ensure its parent permanent
   on disk, to ensure this, cache child locks in mkdir.

Sync-on-Cancel for cross-MDT lock

When two operations have dependency on an object, and the first
operation has a PW/EX cross-MDT lock on this object, trigger
transaction commit on the MDT where the object resides to
eliminate dependency, in short, this patch eliminates dependency
between locks and existed PW/EX cross-MDT lock.

This patch contains following changes:
* enable Sync on Cancel for DNE by default.
* save cross-MDT lock into tgt_uncommitted_soc_locks after use,
  and it will be released upon transaction commit, note, just
  a lock refcount is taken when lock is saved, the read/write
  count is released in mdt_object_unlock().
* the saved cross-MDT lock will be discarded upon BAST,
  because the MDT where the object resides will do sync on lock
  cancel.
* use existed BLOCKING_SYNC_ON_CANCEL mechanism to commit
  transaction upon cross-MDT lock cancel.

Commit-on-Sharing for DNE

On DNE, PW/EX local lock will be saved as COS lock, but such lock
will be ignored in compatilibity check by default, unless it's
required, there are two situations:
1. when distributed transaction locks local object, it will check
   compatibility against COS locks.
2. when distributed transaction enqueues cross-MDT lock, it will
   check compatiblity against COS locks.

This patch contains following changes:
* on DNE, local PW/EX lock is converted to COS and saved like
  before even when COS is not enabled.
* above COS locks will be ignored in lock compatibility check by
  default, so for local operations COS won't take effect. But if
  operation finds itself may modify remote MDT object, it will lock
  all local locks with COS checked.
* cross-MDT lock will always check against COS locks.
* if operation is reint, it will check in each mdt_object_lock()
  to see whether current lock is cross-MDT lock, if so, it will
  relock from start to ensure all local locks are taken with COS
  checked.

Eliminate dependency in striped dir creation

DNE needs to eliminate dependency between 'mkdir a/b' and any
operation on 'a' because 'b' may be a striped directory, which is
able to recover from update log, if 'a' can't be recovered, it
will cause problem.

To achieve this,  mkdir needs to save child lock, so that
subsequent mkdir under it can know the conflict.

Also stripe slave locks will be saved.

Signed-off-by: Lai Siyao <lai.siyao@intel.com>
Change-Id: I91928d097cbb26bd1e1089c3f8851ac6a6440a69