LU-5849 lnet: lnet_parse_networks return number of nets
lnet_dyn_add_ni() requires lnet_parse_networks to return
the number of nets (outside of LO) that were parsed in
order to do proper sanity checks.

lnet_parse_networks() maintained the number of nets parsed
in a local variable nnets, however it didn't retun it.

The change makes lnet_parse_networks() return the number
of nets parsed if successful and < 0 if failure.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: Idcd23f8c1609d50cb0cfa0930ffbcea539d77921