LU-6109 lfsck: NOT purge object by OI scrub
Originally, when the OI scrub found some inconsistent FID mapping,
it will repair the FID mapping and ask others to reload the object
by purging such object. Such behavior is unnecessary and may cause
other (such as LFSCK) to hung. Because if the object corresponding
to the FID has already been established in RAM, then it should be
valid (maybe via lookup FID by name firstly); and if some other
holds the object's reference, such as the LFSCK engine hold the
.lustre/lost+found/MDTxxxx, then such purging object will set
LU_OBJECT_HEARD_BANSHEE on the object, then the subsequent object
find against such FID will be blocked until the object's reference
become zero and re-establish the object in RAM. Unfortunately, if
it is the object's reference holder tries to find the same object,
it will be blocked by itself for ever.

On the other hand, if the OI scrub is triggered by the LFSCK
engine when verify the .lustre/lost+found/ directory, it should
allow the subsequent LFSCK iteration to join.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I651ef5f5e8f4f478f07bcbb5622b345deed7cb31