LU-5851 build: ship osd-related mount DSO out of lustre-osd
Since LU-4606 changes, an osd-specific DSO for mount command is
shipped as part of lustre-osd RPMs. This breaks Lustre Server
DKMS packaging which builds osd module but not this library.

To fix this, now ship the osd-specific DSOs as part of generic utils.

Then to allow this to be functional, had to allow for init() method
of each DSO to fail during load because of missing externals (like
zfs/spl modules for mount_osd_zfs.so, ...).

Test-Parameters: fortestonly
Signed-off-by: Bruno Faccini <bruno.faccini@intel.com>
Change-Id: Ib52c28a16b53be01c03f10d9d868a45058fd3a4b