LU-4009 osd: minimal support for ZIL
in this version all the updates go to ZIL, osd_sync()
and osd_object_sync() flush the whole log.

it's hard to turn all Lustre transactions into a linear sequence
as within a single TXG (actaul commit unit) Lustre puts many
transactions modifying same resources in arbitrary order (so that
transaction T1 can get into ZIL before transaction T2 while T1
depends on T1).

basically ZIL log contains two type of records:
one stores transaction which is a set of updates (see OUT),
another record is a barrier simulating TXG commit, so
we're sure all running transactions have completed by a
barrier and all potential dependent transactions get into
ZIL by this point.

at ZIL replay time we read all the records of the first type
and store them in memory. we also find initial verions for the
objects modified in this set of transactions. once a barrier
met we can start to apply the updates. we do this with regard
to versions. if some update can't applied due to dependency on
the transaction in the future records, we just skip this update
and leave it in memory. in this implementation we just repeat
this process until all the updates are applied.

until all the updates are applied successfully the filesystem
can be called inconsistent. to deal with such a state, we use
snapshots: before ZIL replay start we take a snapshot storing
a consistent state. if the whole replay succeed, then we can
discard the snapshot, otherwise we rollback to that snapshot
and mount fails.

it's a subject to discuss what we should do if ZIL replay fails.

TODO:
 - dnode destroy vs. osd_zil_get_data() race
 - quick search in osd_lock_range()
 - OUT improvements (to make records smaller)
 - improvements to llog (8K writes to llog header)
 - testing (RO should disable ZIL?)
 - benchmark on a real hardware

Q's:
 - how much memory to store all the updates
 - disk space reservation?
 - quota?

Test-Parameters: fortestonly testlist=sanity,sanityn,sanity-benchmark mdtfilesystemtype=zfs mdsfilesystemtype=zfs ostfilesystemtype=zfs alwaysuploadlogs
Change-Id: Ie65dc4dbe47147e0568955c9c8f907cfa84ec701
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>