LU-4647 nodemap: fix problem with node reclassification
nodemap_add_member can't be used to move an already hashed member
to a new nodemap, so this patch copies the needed functionality to
nm_member_reclassif_cb. This also adds a mutex lock for reclassifying
so that there is only one nodemap reclassifying at a time.
Reclassifying takes a lock on a nodemap's nm_member_hash, so a
deadlock could arise if one nodemap is trying to add a member to
another nodemap, and that second nodemap is also reclassifying and
eventually tries to add a member to the first nodemap.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: Icc93a8e6d8384afa90e45cc04f1422512974ce4a